fix all compile errors. no functionality yet (#77)
* fix all compile errors. can't link zlib

* fix oopsie

* nogui still has zlib error. need to unlink in public. dolphin-gui compiles and runs!

* fix cmakelists include

+++ Source/Core/DolphinQt/Settings/SlippiPane.cpp
@@ -92,7 +92,8 @@ void SlippiPane::CreateLayout()
           [](int index) {
             SConfig::GetInstance().m_slippiEnableQuickChat = static_cast<Slippi::Chat>(index);
           });
-  netplay_quick_chat_combo->setCurrentIndex(SConfig::GetInstance().m_slippiEnableQuickChat);
+  netplay_quick_chat_combo->setCurrentIndex(
+      static_cast<u32>(SConfig::GetInstance().m_slippiEnableQuickChat));
 
   online_settings_layout->addRow(tr("Quick Chat:"), netplay_quick_chat_combo);
